+++ packages/harpguru-core/src/components/harp-guru/harp-guru.tsx
@@ -10,19 +10,14 @@ import { LayoutMenu } from '../layout-menu'
 import { HarpFaceMemo } from '../harp-face'
 import { CovariantMenu } from '../covariant-menu'
 import { ActivityLegend } from '../activity-legend'
-import { setGlobalState, setGlobalReducers } from '../../utils'
+import { setGlobalState } from '../../utils'
 import { MenuStates, MenuStashPosition } from '../../types'
 import { colors } from '../../styles'
 import { getSizes } from '../../styles'
 
-import {
-  useFlushBufferedActivityToggles,
-  useMenus,
-  useQuizCycle,
-} from './hooks'
+import { useFlushBufferedActivityToggles, useMenus } from './hooks'
 
 setGlobalState()
-setGlobalReducers()
 
 const styles = StyleSheet.create({
   fillScreen: {
@@ -45,7 +40,6 @@ export const HarpGuru = (): ReactElement => {
   const sizes = getSizes()
   const { 8: swipeThreshold } = sizes
 
-  useQuizCycle(menuState)
   useFlushBufferedActivityToggles()
 
   return (
@@ -74,7 +68,7 @@ export const HarpGuru = (): ReactElement => {
           stashPosition={MenuStashPosition.Bottom}
           openCloseMenu={layoutOpenCloseTapHandler}
         />
-        <QuizQuestionDisplay screenFree={menuState === MenuStates.NoMenu} />
+        <QuizQuestionDisplay isScreenFree={menuState === MenuStates.NoMenu} />
       </View>
     </PanGestureHandler>
   )