+++ packages/harpguru-core/src/components/harp-row/utils/get-harp-cells/get-harp-cells.test.tsx
@@ -14,7 +14,6 @@ mockUseGlobal.mockImplementation((stateItem: string) => {
   if (stateItem === 'activeHarpStrata')
     return [inactiveCellsHarpStrata, jest.fn()]
   if (stateItem === 'activeExperienceMode') return [ExperienceModes.Explore]
-  if (stateItem === 'quizQuestion') return [DegreeIds.Root]
   if (stateItem === 'activeDisplayMode') return [DisplayModes.Degree]
   if (stateItem === 'lockedCovariant') return [CovariantMembers.HarpKey]
   if (stateItem === 'bufferedActivityToggles') return [[]]